pyproject.toml
[project]
name = "heima-agent-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "altair==6.0.0",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "blinker==1.9.0",
    "build==1.4.0",
    "cachetools==6.2.6",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chromadb==1.4.1",
    "click==8.3.1",
    "coloredlogs==15.0.1",
    "cryptography==46.0.3",
    "dashscope==1.25.9",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "durationpy==0.10",
    "filelock==3.20.3",
    "flatbuffers==25.12.19",
    "fastapi==0.116.0",
    "frozenlist==1.8.0",
    "fsspec==2026.1.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.3.4",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jq==1.11.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==35.0.0",
    "langchain==1.2.7",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.7",
    "langchain-ollama==1.0.1",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.7",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.7",
    "langgraph-sdk==0.3.3",
    "langsmith==0.6.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "narwhals==2.16.0",
    "numpy==2.4.1",
    "oauthlib==3.3.1",
    "ollama==0.6.1",
    "onnxruntime==1.23.2",
    "openai==2.15.0",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "orjson==3.11.5",
    "ormsgpack==1.12.2",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.1.0",
    "posthog==5.4.0",
    "propcache==0.4.1",
    "protobuf==6.33.4",
    "pyarrow==23.0.0",
    "pybase64==1.4.3",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pypdf==6.6.2",
    "pypika==0.50.0",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.3.1",
    "rpds-py==0.30.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.46",
    "streamlit==1.54.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.2",
    "toml==0.10.2",
    "tornado==6.5.4",
    "tqdm==4.67.1",
    "typer==0.21.1",
    "typer-slim==0.21.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uuid-utils==0.13.0",
    "uvicorn==0.40.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
